requirements.txt
# BMAD Framework (lightweight implementation)
pydantic>=2.0.0

# Vector Database
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# LLM Clients
ollama>=0.1.0
openai>=1.0.0  # Optional, for OpenAI models

# RAG Utilities
langchain>=0.1.0
langchain-community>=0.0.20

# GUI
customtkinter>=5.2.0
pillow>=10.0.0

# Document Processing
markdown>=3.4.0
python-frontmatter>=1.0.0

# REACH Compliance Integration
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0

# Prop 65 Compliance Integration
openpyxl>=3.1.0
pdfplumber>=0.10.0

# Image Scraping
requests>=2.31.0

# Cloudflare R2 Access
boto3>=1.28.0

# Browser Automation (for advanced scraping)
selenium>=4.15.0

# File Monitoring
watchdog>=3.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0

# Optional: For hybrid search (BM25)
rank-bm25>=0.2.2

# Comprehensive Chatbot Tools Dependencies
# Archive Operations
py7zr>=0.21.0
rarfile>=4.1

# Office Document Operations
python-docx>=1.1.0
python-pptx>=0.6.23

# PDF Operations
pypdf>=3.0.0

# Data Processing
pyyaml>=6.0
toml>=0.10.2

# Text Processing
phonenumbers>=8.13.0
python-dateutil>=2.8.2
chardet>=5.0.0

# System Operations
psutil>=5.9.0

# Image Operations (Pillow already included above)
# PDF to Image conversion (optional, requires poppler)
# pdf2image>=1.16.0

# Media Operations (Video/Audio)
opencv-python>=4.8.0
moviepy>=1.0.3
pydub>=0.25.1
mutagen>=1.47.0

# Book Processing
ebooklib>=0.18
mobi>=0.3
spacy>=3.7.0
isbnlib>=3.10

# Optional: ONNX Runtime for optimized inference
# onnxruntime>=1.16.0
# optimum>=1.14.0

